+++ clang-tools-extra/test/clang-tidy/checkers/Inputs/Headers/cuda/cuda_runtime.h
@@ -1,3 +1,4 @@
 #include "cuda.h"
 
 cudaError_t cudaDeviceReset();
+cudaError_t cudaGetLastError();